ADD sessions controller
